backend/pyproject.toml
[project]
name = "ai-assistant-backend"
version = "0.1.0"
description = "Personal AI assistant backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.20.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
voice = [
    "faster-whisper>=1.0.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
